Split helper.ex into separate files, add mpeg version extraction and inclusion in ADTS structures
